Implement allocate stage for directly upload
The upload process contains some stages. (allocate, upload, tag, register, album)
The allocate stage only exists directly upload to allocate presigned url.
The upload stage is a stage to actually upload a image.
The tag stage is not implemented. It fills image's tags automatically.
The register stage is not implemented. It registers the uploaded image into search pool.
The album stage is a stage to insert album the uploaded image.